requirements.txt
# Telegram Bot
python-telegram-bot[all]>=20.0

# AI
google-genai>=1.51.0

# 数据库
aiosqlite>=0.19.0

# 配置
python-dotenv>=1.0.0

# 工具
Pillow>=10.0.0
aiohttp>=3.9.0
python-dateutil>=2.8.0
paramiko>=3.0.0
feedparser>=6.0.0

# 开发工具（可选）
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0